HACK: Add a TOC0 variant of the fel-sdboot binary
For more information about the TOC0 header format see
    https://linux-sunxi.org/TOC0

This binary has been created with a modified mktoc0.rb
script. It does not include the TOC0 boot code and only
contains the RoT certificate item.

The reason for this is that the TOC0 boot code now needs to
have the execution address, which differs for different SoC
generations (0x0 for H3, 0x10000 for A64/H64/H5) and this
is very inconvenient. But apparently if the TOC0 container
does not include any code item, then the device just drops
into the FEL mode instead of trying some other boot media
according to the boot order (tested on Jide Remix Mini
device with Allwinner H64). And dropping into the FEL mode
is exactly what we want.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>